+++ R/CFObject.R
@@ -59,6 +59,65 @@ CFObject <- R6::R6Class("CFObject",
         cat("\nAttributes:\n")
         print(.slim.data.frame(self$NCvar$attributes, width), right = FALSE, row.names = FALSE)
       }
+    },
+
+    #' @description Add an attribute. If an attribute `name` already exists, it
+    #'   will be overwritten.
+    #' @param name The name of the attribute. The name must begin with a letter
+    #'   and be composed of letters, digits, and underscores, with a maximum
+    #'   length of 255 characters. UTF-8 characters are not supported in
+    #'   attribute names.
+    #' @param type The type of the attribute, as a string value of a netCDF data
+    #'   type.
+    #' @param value The value of the attribute. This can be of any supported
+    #'   type, including a vector or list of values. Matrices, arrays and like
+    #'   compound data structures should be stored as a data variable, not as an
+    #'   attribute and they are thus not allowed. In general, an attribute
+    #'   should be a character value, a numeric value, a logical value, or a
+    #'   short vector or list of any of these. Values passed in a list will be
+    #'   coerced to their common mode.
+    #' @return Self, invisibly.
+    set_attribute = function(name, type, value) {
+      self$NCvar$set_attribute(name, type, value)
+      invisible(self)
+    },
+
+    #' @description Append the text value of an attribute. If an attribute
+    #'   `name` already exists, the `value` will be appended to the existing
+    #'   value of the attribute. If the attribute `name` does not exist it will
+    #'   be created. The attribute must be of "NC_CHAR" or "NC_STRING" type; in
+    #'   the latter case having only a single string value.
+    #' @param name The name of the attribute. The name must begin with a letter
+    #'   and be composed of letters, digits, and underscores, with a maximum
+    #'   length of 255 characters. UTF-8 characters are not supported in
+    #'   attribute names.
+    #' @param value The character value of the attribute to append. This must be
+    #'   a character string.
+    #' @param sep The separator to use. Default is `"; "`.
+    #' @param prepend Logical to flag if the supplied `value` should be placed
+    #'   before the existing value. Default is `FALSE`.
+    #' @return Self, invisibly.
+    append_attribute = function(name, value, sep = "; ", prepend = FALSE) {
+      self$NCvar$append_attribute(name, value, sep, prepend)
+      invisible(self)
+    },
+
+    #' @description Delete an attribute. If an attribute `name` is not present
+    #' this method simply returns.
+    #' @param name The name of the attribute to delete.
+    #' @return Self, invisibly.
+    delete_attribute = function(name) {
+      self$NCvar$delete_attribute(name)
+      invisible(self)
+    },
+
+    #' @description Write the attributes of this object to a netCDF file.
+    #' @param nc The handle to the netCDF file opened for writing.
+    #' @param nm The NC variable name or "NC_GLOBAL" to write the attributes to.
+    #' @return Self, invisibly.
+    write_attributes = function(nc, nm) {
+      self$NCvar$write_attributes(nc, nm)
+      invisible(self)
     }
   ),